compcheck/requirements.txt
anyio==3.7.1
blinker==1.6.2
certifi==2023.7.22
cffi==1.15.1
chardet==5.1.0
click==8.1.6
cryptography==41.0.2
dnspython==2.4.0
exceptiongroup==1.1.2
Flask==2.3.2
future==0.18.3
h11==0.14.0
httpcore==0.17.3
idna==3.4
impacket==0.10.0
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
ldap3==2.9.1
ldapdomaindump==0.9.4
MarkupSafe==2.1.3
pyasn1==0.5.0
pycparser==2.21
pycryptodomex==3.18.0
pyOpenSSL==23.2.0
six==1.16.0
sniffio==1.3.0
Werkzeug==2.3.6
zipp==3.16.2
